run-test.sh: run test cases based on RECIPELIST

+++ run-test.sh
@@ -2,15 +2,14 @@
 
 DEVEL_MODE=
 # LOGLEVEL might be set as an environment variable
-# RECIPEFILES might be set as an environment variable
-# RECIPEDIR might be set as an environment variable
+# RECIPELIST might be set as an environment variable
 # TESTCASE_FILTER might be set as an environment variable
 SHOW_TEST_VERSION=
 # HIGHEST_PRIORITY might be set as an environment variable
 # LOWEST_PRIORITY might be set as an environment variable
 RUN_ALL_WAITING=
 
-while getopts v:s:t:f:Spd:r:DVh:l:w OPT ; do
+while getopts v:s:t:f:SpDVh:l:w OPT ; do
     case $OPT in
         v) export LOGLEVEL="$OPTARG" ;;
         s) KERNEL_SRC="$OPTARG" ;;
@@ -18,8 +17,6 @@ while getopts v:s:t:f:Spd:r:DVh:l:w OPT ; do
         f) TESTCASE_FILTER="$TESTCASE_FILTER $OPTARG" ;;
         S) SCRIPT=true ;;
 		p) SUBPROCESS=true ;;
-		d) RECIPEDIR="$OPTARG" ;;
-		r) RECIPEFILES="$RECIPEFILES $OPTARG" ;;
 		D) DEVEL_MODE=true ;;
 		V) SHOW_TEST_VERSION=true ;;
 		h) HIGHEST_PRIORITY=$OPTARG ;;
@@ -46,27 +43,6 @@ if [ "$SHOW_TEST_VERSION" ] ; then
 	exit 0
 fi
 
-for rd in $RECIPEDIR ; do
-	if [ -d "$rd" ] ; then
-		for rf in $(find $(readlink -f $rd) -type f) ; do
-			RECIPEFILES="$RECIPEFILES $rf"
-		done
-	fi
-done
-
-make --no-print-directory allrecipes | grep ^cases > $GTMPD/full_recipe_list
-make --no-print-directory RUNNAME=$RUNNAME waiting_recipes | grep ^cases > $GTMPD/waiting_recipe_list
-
-if [ ! "$RECIPEFILES" ] ; then
-	if [ "$RUN_ALL_WAITING" ] ; then
-		RECIPEFILES="$(cat $GTMPD/waiting_recipe_list)"
-	else
-		echo "RECIPEFILES not given or not exist." >&2
-		exit 1
-	fi
-fi
-export RECIPEFILES="$(readlink -f $RECIPEFILES)"
-
 . $TCDIR/lib/recipe.sh
 . $TCDIR/lib/patch.sh
 . $TCDIR/lib/common.sh
@@ -79,10 +55,6 @@ stop_test_running() {
 }
 
 trap stop_test_running SIGTERM SIGINT
-
-echo_log "=========> start testing $(basename $TRDIR):$TESTNAME"
-echo_log "RECIPEFILES:"
-echo_log "${RECIPEFILES//$TRDIR\/cases\//}"
 
 echo 1 > /proc/sys/kernel/panic_on_oops
 echo 1 > /proc/sys/kernel/softlockup_panic
@@ -101,49 +73,43 @@ run_recipe() {
 	export TMPF=$TMPD
 	export OFILE=$TMPD/result
 
-	if check_testcase_already_run ; then
-		echo_log "### You already have workfiles for recipe $recipe_relpath with TESTNAME: $TESTNAME, so skipped. If you really want to run with removing old work directory, please give environment variable AGAIN=true."
-		return
-	fi
-
-	check_remove_suffix $recipe || return
-
 	parse_recipefile $recipe .tmp.recipe
 
-	if [ -d $TMPD ] ; then
+	if [ -d $TMPD ] && [ "$AGAIN" == true ] ; then
 		rm -rf $TMPD/* > /dev/null 2>&1
-	else
-		mkdir -p $TMPD > /dev/null 2>&1
 	fi
-
-	date +%s%3N > $TMPD/start_time
-	# TODO: put general system information under $TMPD
-	# prepare empty testcount file at first because it's used to check
-	# testcase result from summary script.
-	reset_per_testcase_counters
-	init_return_code
-
-	PRIORITY=10 # TODO: better place?
+	mkdir -p $TMPD > /dev/null 2>&1
 
 	mv .tmp.recipe $TMPD/_recipe
+	PRIORITY=10 # TODO: better place?
+	. $TMPD/_recipe
+	ret=$?
 	echo_log "===> testcase '$TEST_TITLE' start" | tee /dev/kmsg
-	(
-		. $TMPD/_recipe
-		ret=$?
-		if [ "$SKIP_THIS_TEST" ] ; then
-			echo_log "This testcase is marked to be skipped by developer."
-			echo_log "TESTCASE_RESULT: $recipe_relpath: SKIP"
-		elif [ "$ret" -ne 0 ] ; then
-			echo_log "TESTCASE_RESULT: $recipe_relpath: SKIP"
-		elif [ "$PRIORITY" ] && [ "$HIGHEST_PRIORITY" -gt "$PRIORITY" ] ; then
-			skip_testcase_out_priority
-		elif [ "$PRIORITY" ] && [ "$LOWEST_PRIORITY" -lt "$PRIORITY" ] ; then
-			skip_testcase_out_priority
-		else
-			do_soft_try
-		fi
-	) 2>&1 | tee -a $OFILE
-	date +%s%3N > $TMPD/end_time
+
+	if check_testcase_already_run ; then
+		echo_log "### You already have workfiles for recipe $recipe_relpath with TESTNAME: $TESTNAME, so skipped. If you really want to run with removing old work directory, please give environment variable AGAIN=true."
+	elif ! check_remove_suffix $recipe ; then
+		return
+	elif [ "$SKIP_THIS_TEST" ] ; then
+		echo_log "This testcase is marked to be skipped by developer."
+		echo_log "TESTCASE_RESULT: $recipe_relpath: SKIP"
+	elif [ "$ret" -ne 0 ] ; then
+		echo_log "TESTCASE_RESULT: $recipe_relpath: SKIP"
+	elif [ "$PRIORITY" ] && [ "$HIGHEST_PRIORITY" -gt "$PRIORITY" ] ; then
+		skip_testcase_out_priority
+	elif [ "$PRIORITY" ] && [ "$LOWEST_PRIORITY" -lt "$PRIORITY" ] ; then
+		skip_testcase_out_priority
+	else
+		date +%s%3N > $TMPD/start_time
+		# TODO: put general system information under $TMPD
+		# prepare empty testcount file at first because it's used to check
+		# testcase result from summary script.
+		reset_per_testcase_counters
+		init_return_code
+
+		( do_soft_try ) 2>&1 | tee -a $OFILE
+		date +%s%3N > $TMPD/end_time
+	fi
 	echo_log "<=== testcase '$TEST_TITLE' end" | tee /dev/kmsg
 }
 
@@ -151,7 +117,6 @@ run_recipe_tree() {
 	local dir="$1"
 
 	if [ -f "$dir/config" ] ; then
-		echo_log "sourcing $dir/config"
 		. "$dir/config"
 	fi
 
@@ -160,7 +125,6 @@ run_recipe_tree() {
 			if [ "$f" == config ] ; then
 				true
 			elif [ -f "$dir/$f" ] ; then
-				echo "$BASHPID f $dir/$f VAL=$VAL"
 				run_recipe "$dir/$f"
 			elif [ -d "$dir/$f" ] ; then
 				run_recipe_tree "$dir/$f"
@@ -174,7 +138,7 @@ get_next_level() {
 	local dir="$1"
 	local list="$2"
 
-	grep ^$dir/ $list | sed -e "s|^$dir/||" | cut -f1 -d / | uniq
+	cat $list | sed "s|^|$PWD/|" | grep ^$dir/ | sed -e "s|^$dir/||" | cut -f1 -d / | uniq
 }
 
 run_recipe_list() {
@@ -182,7 +146,6 @@ run_recipe_list() {
 	local list="$2"
 
 	if [ -f "$dir/config" ] ; then
-		echo_log "sourcing $dir/config"
 		. "$dir/config"
 	fi
 
@@ -191,7 +154,6 @@ run_recipe_list() {
 			if [ "$f" == config ] ; then
 				true
 			elif [ -f "$dir/$f" ] ; then
-				echo "$BASHPID f $dir/$f VAL=$VAL"
 				run_recipe "$dir/$f"
 			elif [ -d "$dir/$f" ] ; then
 				run_recipe_list "$dir/$f" "$list"
@@ -212,5 +174,11 @@ run_recipes() {
 	fi
 }
 
-echo run_recipes $TRDIR/cases
-run_recipes $TRDIR/cases
+if [ -f "$RECIPELIST" ] ; then
+	cp $RECIPELIST $GTMPD/recipelist
+elif [ ! -f "$GTMPD/recipelist" ] ; then
+	make --no-print-directory allrecipes | grep ^cases > $GTMPD/recipelist
+fi
+# make --no-print-directory RUNNAME=$RUNNAME waiting_recipes | grep ^cases > $GTMPD/waiting_recipe_list
+
+run_recipes $TRDIR $GTMPD/recipelist
